+++ src/questFolder/clearTile.py
@@ -3,7 +3,7 @@ import random
 import src
 
 
-class ClearTile(src.quests.MetaQuestSequence):
+class ClearTile(src.quests.MetaQuestSequenceV2):
     type = "ClearTile"
 
     def __init__(self, description="clean tile", creator=None, targetPosition=None, noDelegate=False, reason=None):
@@ -76,27 +76,6 @@ Remove all items from the walkways that are not bolted down."""
             self.targetPosition = parameters["targetPosition"]
             self.metaDescription = self.baseDescription+" "+str(self.targetPosition)
         return super().setParameters(parameters)
-
-    def solver(self, character):
-        if self.triggerCompletionCheck(character):
-            return
-
-        (nextQuests,nextCommand) = self.getNextStep(character)
-        if nextQuests:
-            for quest in nextQuests:
-                self.addQuest(quest)
-            return
-
-        if nextCommand:
-            character.runCommandString(nextCommand[0])
-            return
-        super().solver(character)
-
-    def getSolvingCommandString(self, character, dryRun=True):
-        nextStep = self.getNextStep(character)
-        if nextStep == (None,None):
-            return super().getSolvingCommandString(character)
-        return self.getNextStep(character)[1]
 
     def getNextStep(self,character=None,ignoreCommands=False):
         if self.subQuests:
